feat : 슬라이더 스켈레톤, 파일명 등 수정
Related to : #139
