--- src/styles/main/mainSlider/customArrow.module.css
@@ -1,46 +1,0 @@
-.arrow {
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	position: absolute;
-	width: 40px;
-	height: 40px;
-	border-radius: 50%;
-	box-sizing: border-box;
-	background-color: #ffffff;
-	z-index: 10;
-	transition:
-		width 0.2s ease,
-		height 0.2s ease,
-		left 0.2s ease,
-		right 0.2s ease;
-}
-
-.arrowPrev {
-	left: 180px;
-}
-
-.arrowNext {
-	right: 180px;
-}
-
-.arrow:hover {
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	position: absolute;
-	width: 47px;
-	height: 47px;
-	border-radius: 50%;
-	box-sizing: border-box;
-	background-color: #ffffff;
-	z-index: 1;
-}
-
-.arrowPrev:hover {
-	left: calc(180px - ((45px - 38px) / 2));
-}
-
-.arrowNext:hover {
-	right: calc(180px - ((45px - 38px) / 2));
-}
