--- src/styles/main/mainSlider/mainSlider.module.css
@@ -1,9 +1,0 @@
-.mainSliderWrapper {
-	width: 1300px;
-	height: 400px;
-	border-radius: 10px;
-
-	display: flex;
-	align-items: center;
-	justify-content: center;
-}
