--- src/styles/main/mainSlider/sliderItem.module.css
@@ -1,64 +1,0 @@
-.sliderItem {
-	/* width: 800px; */
-	height: 400px;
-	border-radius: 10px;
-	margin: 0 4px;
-
-	display: flex;
-	align-items: center;
-	justify-content: center;
-}
-.sliderImg {
-	border-radius: 5px;
-	position: relative;
-	box-shadow: -20px 16px 9px 8px rgba(0, 0, 0, 0.1);
-	margin-left: 20px;
-}
-.sliderTitleWrap {
-	position: absolute;
-	padding: 40px;
-	bottom: 43px;
-	text-align: left;
-}
-.sliderCategory {
-	opacity: 1;
-	color: #fff;
-	height: 27px;
-	line-height: 27px;
-	font-size: 18px;
-	font-weight: 400;
-}
-.sliderTitle {
-	opacity: 1;
-	color: #fff;
-	margin-top: 8px;
-	max-height: 68px;
-	overflow: hidden;
-	line-height: 34px;
-	font-size: 30px;
-	font-weight: 700;
-}
-.sliderTitle2 {
-	opacity: 1;
-	color: #fff;
-	max-height: 68px;
-	overflow: hidden;
-	line-height: 34px;
-	font-size: 30px;
-	font-weight: 700;
-}
-.sliderContent {
-	opacity: 1;
-	color: rgba(255, 255, 255, 0.6);
-	margin-top: 7px;
-	height: 24px;
-	line-height: 24px;
-	font-size: 16px;
-	font-weight: 500;
-}
-.center .sliderTitle2 {
-	opacity: 0;
-}
-/* .customSliderItem.slick-center .sliderTitle2 {
-	opacity: 1 !important;
-} */
