+++ src/styles/main/slider/slider.module.css
@@ -1,0 +1,9 @@
+.mainSliderWrapper {
+	width: 1300px;
+	height: 400px;
+	border-radius: 10px;
+
+	display: flex;
+	align-items: center;
+	justify-content: center;
+}
