+++ src/styles/not-found.module.css
@@ -24,3 +24,37 @@
 .btnWrap :nth-child(1) {
 	margin-right: 50px;
 }
+
+.container {
+	height: calc(100vh - 290px);
+	gap: 50px;
+
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	flex-direction: column;
+}
+.notFountWrapper {
+	width: 400px;
+	font-size: 20px;
+	gap: 50px;
+
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+}
+.notFountWrapper > div {
+	font-size: 14px;
+	gap: 50px;
+
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	flex-direction: row;
+}
+.notFountWrapper > div > div {
+	cursor: pointer;
+}
+.notFountWrapper > div > a {
+	color: #000000;
+}
